Modularisation of test setup
