+++ gulpfile.js
@@ -2,7 +2,7 @@ var gulp = require('gulp'),
     mocha = require('gulp-mocha');
 
 gulp.task('test', function() {
-    return gulp.src(['test/**.js'], {
+    return gulp.src(['test/allTests.js'], {
             read: false
         })
         .pipe(mocha({
