+++ test/allTests.js
@@ -1,0 +1,17 @@
+function importTest(name, path) {
+    describe(name, function() {
+        require(path);
+    });
+}
+
+describe("Running all tests ...", function() {
+    importTest("Uber client general tests", './general');
+    importTest("OAuth2 authorization methods", './oauth');
+    importTest("/Estimates", './estimates');
+    importTest("/Payment-Methods", './payment-methods');
+    importTest("/Places", './places');
+    importTest("/Products", './products');
+    importTest("/Reminders", './reminders');
+    importTest("/User", './user');
+    importTest("Deprecated methods", './deprecated');
+});
